scripts/requirements.txt
spotipy>=2.23.0
pytrends>=4.9.0
google-api-python-client>=2.100.0
requests>=2.31.0
python-dotenv>=1.0.0
